requirements.txt
opencv-python
face_recognition
numpy
pyserial
